Define config files using individual fragments
This commit should not change the overall behaviour of the module
for existing users. However, it will require either ripienaar-concat
or https://github.com/Savar/file_concat.  If the global variable
*bacula_use_file_concat* is set to +true+ then the latter module will
be used.

To define an arbitrary fragment for bacula-dir.conf, use the
*bacula::director::config_fragment* definition.  For example:

   bacula::director::config_fragment { 'my-job':
     content => template('site/bacula/my-job.erb')
   }

For some configuration objects there are specialized definitions that
should be used in favor of arbitrary fragments to increaes consistency
and allow for parameter overrides when exported resources are realized,
for example.

The same pattern applies to the storage and the client configuration.